UPDATE: X1MK3 - Community Performance MOD Version 11 beta 4f TP 4.2.0 Remix Deck
Hold SHIFT and turn Loop Encoder to adjust Capture Source.
Hold SHIFT and push Loop Encoder to toggle Quantization.
Hold SHIFT and hold down Loop Encoder and turn Loop Encoder to adjust Quantization (and automatically engage as well).
Requirements

In LegacyEncoder mode, set SHIFT + Loop Encoder to Key-adjust. Setting it to Beatjump/Move will override the new controls.
In BrowserMode it only works if browsing is not active; the browsing controls override it (of course).
Note: No screen display for this proof of concept. This is for you to test
https://community.native-instruments.com/discussion/comment/229548#Comment_229548