UPDATE: X1MK3 - Community Performance MOD Version 11 beta 4h TP 4.2.0 :
Screen feedback complete. I had to revert the beta4g changes back to beta4f so that there was no confusion wether or not Quantize is active (Quant Index is black on white when active).

Notes:

Hotcues 1-4 (and on SHIFT layer: delete Hotcues 1-4) need to be selected (X1MK3 pane in Preferences) for the number buttons! They are used for remix slot functions.
The (Hotcue→RemixSlot) functions are LOAD (when empty: from capture source) / PLAY (when loaded and stopped) / (UN)MUTE (when loaded and playing).
The SHIFT (delete Hotcue) functions are LOAD (when empty: from playlist) / STOP (when loaded and playing) / DELETE (when loaded and stopped)
http://community.native-instruments.com/discussion/comment/229566#Comment_229566